packages/rule-engine/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rule-engine"
version = "0.1.0"
description = "Threatflow rule engine for OTM models with pluggable analyzers"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2",
  "jmespath>=1.0.1",
  "otm-model>=0.1.0",
  "pyyaml>=6",
]

[project.optional-dependencies]
threagile = []

[tool.setuptools.packages.find]
where = ["src"]

